moving menu item to shared component
